requirements.txt
asgiref==3.7.2
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2==2.9.6
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
